Complete excercises 4.21-4.22 and Finish part 4

+++ controllers/blogs.js
@@ -2,7 +2,6 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
-const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -28,6 +27,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if(!blog) {
+        return response.status(404).json({ error: 'blog with that id does not exist' })
+    }
+        
+    if(decodedToken.id.toString() !== blog.user.toString()) {
+        return response.status(401).json({ error: 'operation not permitted' })
+    }
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
